fix failing test in contacts app, moved custom validators
